[SlateEditor] update implementation, remove styled-components package
