+++ uui-editor/package.json
@@ -46,12 +46,10 @@
     "slate-react": "^0.82.2",
     "slate-soft-break": "^0.9.0",
     "slate-uui-table-plugin": "^1.0.6",
-    "styled-components": "5.3.1",
     "tippy.js": "^6.3.7",
     "use-force-update": "^1.0.11"
   },
   "devDependencies": {
-    "@types/styled-components": "5.1.25",
     "react-popper": "2.3.0",
     "slate": "0.47.9",
     "slate-drop-or-paste-images": "0.9.1",
