+++ uui-editor/src/plate/plugins/tablePlugin/Table.scss
@@ -54,23 +54,75 @@
     }
 }
 
-.cell {
-    border: 1px solid $gray30;
-    position: relative;
-
-    & > :first-child {
-        // padding: 0; leads to bugs, see link below
-        // https://stackoverflow.com/questions/27829250/why-doesnt-a-childs-margin-affect-a-parents-height
-        padding: 6px 24px;
-
-        & > p {
-            padding: 0;
-        }
-    }
-}
-
 .cells-selection-active {
     *::selection {
         background: none;
     }
+}
+
+.table-cell-wrapper {
+    --tw-bg-opacity: 1;
+    position: relative;
+    overflow: visible;
+    background-color: rgb(255 255 255 / var(--tw-bg-opacity));
+    padding: 0;
+    text-align: left;
+
+    &::before {
+        content: '';
+        position: absolute;
+        width: 100%;
+        height: 100%;
+        box-sizing: border-box;
+        user-select: none;
+        border-bottom: 1px solid rgb(209 213 219);
+        border-right: 1px solid rgb(209 213 219);
+    }
+}
+
+.table-cell-wrapper.table-cell-border-left::before {
+    border-left: 1px solid rgb(209 213 219);
+}
+
+.table-cell-wrapper.table-cell-border-top::before {
+    border-top: 1px solid rgb(209 213 219);
+}
+
+.table-cell-wrapper.table-cell-selected::before {
+    --tw-bg-opacity: 1;
+    background-color: rgb(239 246 255 / var(--tw-bg-opacity));
+}
+
+.table-cell-content {
+    // "padding: 0;" leads to bugs, see link below
+    // https://stackoverflow.com/questions/27829250/why-doesnt-a-childs-margin-affect-a-parents-height
+    padding: 6px 24px;
+    position: relative;
+    z-index: 20;
+    box-sizing: border-box;
+    height: 100%;
+
+    &>p {
+        padding: 0;
+    }
+}
+
+@mixin resize-handle {
+    --tw-bg-opacity: 1;
+    position: absolute;
+    z-index: 30;
+    width: 0.25rem;
+    background-color: rgb(59 130 246 / var(--tw-bg-opacity));
+    top: 0;
+    height: 100%;
+}
+
+.table-cell-resize-right-handle {
+    @include resize-handle();
+    right: -1.5px;
+}
+
+.table-cell-resize-left-handle {
+    @include resize-handle();
+    left: -1.5px;
 }
